+++ AcadeMeData/tests.py
@@ -1,5 +1,5 @@
 import pytest
-from AcadeMeData.models import User, Professor, University, Degree
+from AcadeMeData.models import User, Professor, University
 
 
 @pytest.mark.django_db
@@ -88,22 +88,3 @@ class TestProfessorModel:
         professor = Professor.get_professor(name)
         assert professor.get_name() == professor_for_example.name
         assert professor.get_description() == professor_for_example.description
-
-
-@pytest.mark.django_db
-class TestDegreeModel:
-    @pytest.fixture
-    def generate_degree(self, degree_id=1, name='History', universities="Ben Gurion University, Reichman University",
-                        description="Learn about historic events and their influences on the world"):
-        degree = Degree(degree_id=degree_id, name=name, universities=universities, description=description)
-        degree.save()
-        return degree
-
-    def test_create_degree(self):
-        degree = Degree.create_degree(degree_id=1, name="History",
-                                      universities="Reichman University, Ben Gurion University",
-                                      description="Learn about historic events and their influences on the world")
-
-        assert "Reichman University" in degree.universities
-        assert "History" in degree.name
-        assert "historic" in degree.description
